par01

+++ src/stack/stack.h
@@ -8,7 +8,7 @@
 #include "errors.h"
 
 typedef struct stack {
-    ssize_t *data;
+    double *data;
     ssize_t top;
     ssize_t size;
 } stack_t;
@@ -22,11 +22,11 @@ bool stack_is_empty(stack_t *stack);
 bool stack_is_full(stack_t *stack);
 status_t extend_stack(stack_t *stack);
 
-status_t push(stack_t *stack, ssize_t value);
+status_t push(stack_t *stack, double value);
 
-ssize_t pop(stack_t *stack);
+double pop(stack_t *stack);
 
-ssize_t stack_peek(stack_t *stack);
+double stack_peek(stack_t *stack);
 
 void stack_print(stack_t *stack);
